grind out some object/map stuff
