+++ include/game/models/model.h
@@ -1,0 +1,36 @@
+#include "PR/ultratypes.h"
+
+typedef struct Model {
+    //XXX most fields need to be filled in.
+    //just padding them out for now.
+    s32 unk00;
+    s32 unk04;
+    s32 unk08;
+    s32 unk0C;
+    s32 unk10;
+    s32 unk14;
+    s32 unk18;
+    s32 unk1C;
+    s32 unk20;
+    s32 unk24;
+    void **animations;
+    s32 unk2C;
+    s32 unk30;
+    s32 unk34;
+    s32 unk38;
+    s32 unk3C;
+    s32 unk40;
+    s32 unk44;
+    s32 unk48;
+    s32 unk4C;
+    s32 unk50;
+    s32 unk54;
+    s32 unk58;
+    s32 unk5C;
+    s32 unk60;
+    s16 unk64;
+    s16 unk66;
+    s32 unk68;
+    s32 unk6C;
+    //probably more...
+} Model;
